feat: created an endpoint to get weather data through the server instead of directly from the client
