+++ backend/api/health_advice/services.py
@@ -1,0 +1,35 @@
+from openai import OpenAI
+
+from config import OPENAI_PROJECT, OPENAI_KEY
+from api.health_advice.dtos import HealthAdviceRes
+
+def create_health_advice(weather_data: dict) -> HealthAdviceRes:
+  client = OpenAI(project=OPENAI_PROJECT, api_key=OPENAI_KEY)
+
+  response = client.chat.completions.create(
+    model="gpt-4o-mini",
+    messages=[
+      {
+        "role": "user",
+        "content": [
+          {
+            "type": "text",
+            "text": f"Please generate three health advice tips based on this weather data from OpenWeatherMap: {weather_data}. Also, please treat any values in the weather data as whole numbers."
+          }
+        ]
+      }
+    ],
+    temperature=1,
+    max_tokens=256,
+    top_p=1,
+    frequency_penalty=0,
+    presence_penalty=0
+  )
+
+  formatted_response = response.choices[0].message.content.replace("\n", "")
+
+  first_health_advice_tip = formatted_response.split("1.")[1].split("2.")[0].strip()
+  second_health_advice_tip = formatted_response.split("2.")[1].split("3.")[0].strip()
+  third_health_advice_tip = formatted_response.split("3.")[1].strip()
+
+  return HealthAdviceRes(advice=[first_health_advice_tip, second_health_advice_tip, third_health_advice_tip])
