+++ backend/api/weather/dtos.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class WeatherRes(BaseModel):
+  msg: str
+  weather_data: dict
