+++ backend/config.py
@@ -7,3 +7,4 @@ load_dotenv()
 SECRET_KEY = getenv("SECRET_KEY")
 OPENAI_KEY = getenv("OPENAI_KEY")
 OPENAI_PROJECT = getenv("OPENAI_PROJECT_ID")
+OPENWEATHERMAP_KEY = getenv("OPENWEATHERMAP_KEY")
